Some profound changes: Added leader move as argument to get_move and Made trump exchange a proper Trick, creating an additional state.
The get_move method had an issue that the follower had to somehow get the previous partial trick from the state, polluting the PlayerGameState with partial information about a trick, which is not relevant for later history lookups, etc.
To clear that out, the previous trick is not puv=blicly part of the PlayerGameState and instead given when the move is requested.

A trump exchange used to piggyback on a normal trick. The problem was that this made the code rather convoluted.
The issue was that the trump exchange had to be applied before the leader (and follower) do their actual moves, but at that point we are in limbo between the old state and a partially modified new state.
To clear this up, the trump exchange has been promoted to a full Tirck, meaning that when a trump exchange is played, a new GameState is created and for this Trick the second player is not asked for a move, not is any scoring done.

The testing code has been adapted accordingly, but quite a few cases are not yet covered in the test. More bugs will still be found.

Trigger CI