This commit is about the implementation of get_game_history.
Implemented get_game_history in PlayerGameState.
To do this, moved the previous information out of GameState into its own object, which also keeps track on changes in the leadership.
The Concrete PlayerState classes now also implement am_i_leader()->bool which indicates whether we are leading or following.
Added a historybot to the CLI to test this out

Trigger CI